+++ roko/datasets.py
@@ -135,5 +135,5 @@ class InMemoryTrainDataset(Dataset):
 class TrainToTensor:
     def __call__(self, sample):
         X, Y, X2, X3 = sample
-        X2 = X2.astype(np.int16)
+        #X2 = X2.astype(np.int16)
         return torch.from_numpy(X), torch.from_numpy(Y), torch.from_numpy(X2), torch.from_numpy(X3)
